% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{T2DMantidiabetic-package}
\alias{T2DMantidiabetic}
\alias{T2DMantidiabetic-package}
\title{T2DMantidiabetic: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
